+++ iOS/RELEASE_NOTES.md
@@ -31,3 +31,4 @@
 * Prevent crashing in a rare situation where you manage to return to home while the app is already returning to home
 * Fail gracefully rather than crashing when saving a corrupt image that cannot be converted into a png (This appears in a crash report, but we don't know the conditions where this actually happens)
 * Fail gracefully rather than crashing when failing to save a file to the Files App (This appears in a crash report, but we don't know the conditions where this actually happens; perhaps no disk space?)
+* Prevent crashing when linking to the LeafByte website on very slow internet